requirements.txt
click

pyyaml

Jinja2

markdown


